blktrace: change size of action to 64 bits
In order to add the zoned commands to blktrace's actions, the storage size
needs to be increased to 64bits.

In order to keep the structure alignment 'action' and 'pid' need to be
swapped as well.

Therefore blktrace's version number also needs to be bumped.

+++ include/linux/blktrace_api.h
@@ -18,7 +18,7 @@ struct blk_trace {
 	struct rchan *rchan;
 	unsigned long __percpu *sequence;
 	unsigned char __percpu *msg_data;
-	u16 act_mask;
+	u64 act_mask;
 	u64 start_lba;
 	u64 end_lba;
 	u32 pid;
@@ -99,7 +99,7 @@ static inline int blk_trace_remove(struct request_queue *q)
 
 struct compat_blk_user_trace_setup {
 	char name[BLKTRACE_BDEV_SIZE];
-	u16 act_mask;
+	u32 act_mask;
 	u32 buf_size;
 	u32 buf_nr;
 	compat_u64 start_lba;
